fix: update expected type

+++ systemtest/metrics_test.go
@@ -170,7 +170,7 @@ func TestApplicationMetrics(t *testing.T) {
 			"full_name": "a.b.c",
 			"mapping": map[string]interface{}{
 				"c": map[string]interface{}{
-					"type": "float",
+					"type": "double",
 				},
 			},
 		},
@@ -178,7 +178,7 @@ func TestApplicationMetrics(t *testing.T) {
 			"full_name": "x.y.z",
 			"mapping": map[string]interface{}{
 				"z": map[string]interface{}{
-					"type": "float",
+					"type": "double",
 				},
 			},
 		},
